remove xorg.conf mention (superseded by xorg.conf.d snippet, or HAL on older systems)
The xorg.conf option is not as straightforward as I thought:
Since hotplugging with HAL, the Input Sections of xorg.conf are
automatically ignored. When adding an input section, one would have to
disable hotplugging with Option "AutoAddDevices" "false".
But it is a lot simpler to just write a HAL policy file in that case.