Refactor structs XYinfo, AxisInfo and clicked
AxisInfo contains info on a single axis, including inversion. XYinfo
contains an AxisInfo for X and Y, as well as swap_xy.
Clicked are no longer 2 arrays but a struct with vectors and the count.

This patch was originally part of Antoine Hue's commit with message
'''Managing correctly X/Y swap and X or Y inversion: - detect swap and
inversion modification - take into account for inversion in calibration
computation since Evdev is doing inversion after calibration. - Mainly
tested for Evdev based driver. To be tested for USB.'''

Signed-off-by: Tias Guns <tias@ulyssis.org>