Tester: test calibration routines and make assumptions explicit
The tester will serve to test the calibration code,
while doing so, it will make the assumptions about the calibration
in the driver (e.g. how it uses the calibration values, and
especially swap_xy and inversion and the like) explicit.

The goal is to give assurances about the correctness of the calibration,
without needing to have devices in all possible swapaxes/inversion
combinations.