Add option for setting timeout value
This commit adds a '--timeout <N>' option by which users can specify the
amount of seconds for the timeout, or 0 to disable it.

This is useful in those cases in which the default value (15 seconds) is
too short or even too long.

Actually the existing '--no-timeout' has been kept for compatibility and
is a shortcut for '--timeout 0'.